Внесены изменения по замечаниям
