+++ src/main/java/pro/sky/employee/DepartmentService.java
@@ -3,11 +3,11 @@ package pro.sky.employee;
 import java.util.List;
 import java.util.Map;
 
-public interface EmployeeBookService {
+public interface DepartmentService {
     Employee employeeWithMaxSalaryInDepartment(int depart);
     Employee employeeWithMinSalaryInDepartment(int depart);
     List<Employee> printEmployeeInDepartment(int depart);
-    Map<String,Employee> getAllEmployeeWithDepartment();
+    Map<Integer,List<Employee>> getAllEmployeeWithDepartment();
 
 
 }
